Feeshare: Move to new params + test + lint
